Flaggy'sNotebook/challenge/requirements.txt
requests
flask
svglib==0.9.3
reportlab
pillow
flask-login
werkzeug
